+++ dev/src/hasher.js
@@ -14,6 +14,8 @@ var hasher = (function(window){
 
     var
 
+        // frequency that it will check hash value on IE 6-7 since it doesn't
+        // support the hashchange event
         POOL_INTERVAL = 25,
 
         // local storage for brevity and better compression --------------------------------
@@ -36,10 +38,16 @@ var hasher = (function(window){
 
         // sniffing/feature detection -------------------------------------------------------
 
-        _isIE = (!+"\v1"), //hack based on this: http://webreflection.blogspot.com/2009/01/32-bytes-to-know-if-your-browser-is-ie.html
-        _isHashChangeSupported = ('onhashchange' in window), // FF3.6+, IE8+, Chrome 5+, Safari 5+
-        _isLegacyIE = _isIE && !_isHashChangeSupported, //check if is IE6-7 since hash change is only supported on IE8+ and changing hash value on IE6-7 doesn't generate history record.
-        _isLocal = (window.location.protocol === 'file:');
+        //hack based on this: http://webreflection.blogspot.com/2009/01/32-bytes-to-know-if-your-browser-is-ie.html
+        _isIE = (!+"\v1"),
+        // hashchange is supported by FF3.6+, IE8+, Chrome 5+, Safari 5+ but
+        // feature detection fails on IE compatibility mode, so we need to
+        // check documentMode
+        _isHashChangeSupported = ('onhashchange' in window) && document.documentMode !== 7,
+        //check if is IE6-7 since hash change is only supported on IE8+ and
+        //changing hash value on IE6-7 doesn't generate history record.
+        _isLegacyIE = _isIE && !_isHashChangeSupported,
+        _isLocal = (location.protocol === 'file:');
 
 
     //--------------------------------------------------------------------------------------